Deployment/requirements.txt
numpy==1.21.6
tensorflow==2.8.0
opencv-python==4.1.2.30
cvlib==0.2.7
pandas==1.3.5
streamlit==1.9.0
joblib==1.1.0
pillow==7.1.2
opencv-contrib-python==4.1.2.30
opencv-python-headless==4.5.3.56
matplotlib==3.2.2
